Added a function that handles the ChargeMessage when trying to charge incorrectly

+++ User_WalletServices_Tests/Models/WalletServiceChargeRequest.cs
@@ -3,7 +3,7 @@
     public class WalletServiceChargeRequest
     {
         public int? userId;
-        public decimal? amount;
+        public double? amount;
 
     }
 }
